requirements.txt
# Core
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Biomedical signal processing
mne>=1.5.0
neurokit2>=0.2.7
wfdb>=4.1.0
antropy>=0.1.6
PyWavelets>=1.4.1

# ML / DL extras
lightgbm>=4.1.0
shap>=0.43.0
captum>=0.7.0

# Federated learning
syft>=0.8.0b0

# Real-time & streaming
pylsl>=1.16.0
brainflow>=5.10.0

# Visualization / Dashboard
streamlit>=1.28.0
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0
rich>=13.6.0
click>=8.1.0
loguru>=0.7.2
joblib>=1.3.0
h5py>=3.9.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
